Added ability to animate any framework element through the use of attached properties (similar to the way pages can be animated).
Popup Window now animates in from the bottom and displays the change in the number of records in the database since the last acknowledgement was received.

Added debug tools to be able to zero out this Acknowledgement counter in DBMonitor.cs

Added a dimmed overlay border with visibility tied to the popup visibility.  If the popup is visible, the opacity mask of 60% is applied over the UI controls (menus and status bar and debugging tools are excluded from this "lockout")

Additional minor cosmetic changes have been added.