Significant restructuring.  Created a Pharmay2U Core library for housing common functionality that is used by all projects in the application.  Probably need to create something similar for the animation, font, and style items items since they can be used by multiple applications as well within the project
Moved the DatabaseChangedControl to the PopupMonitor application.  BIndings and links do not work at this time.

Cleared errors (whichappears to have  solved the issue with the Application design mode in Visual Studio now working.

Updated configurations for debug and release.

Pipe information is still included but needs to be reinstalled for both the application and the monitor.  Retesting of connections by both applications to the pop needs to be verified.