Separated the styles into a separate class library (*.dll) projects within the solution.  Will do the same for Animations so that all projects in the solution will have access to the same code.
Imported official Deaconess Pharmacy 2U logos.

Created the Full Information Order form with placeholders for the field values.  TODO:  Need to apply bindings to this form and to link this form to the summary element in the SQL show pages..

Wrestled with an issue with the scroll viewer -- Note to self:  Remember that StackPanels cause crazy issues with ScrollViewer elements because they can size to anything that they want.  Always place elements in grids whenever scrollviewers may be needed.