Refactored page animations and framework animations to a new class library Pharm2UAnimation.  This will facilitate using the same code in multiple projects, as well as starts the process of using DLL files for future, non Pharm2U projects as they arise?
Made a new BaseAnimationAttachedProperty (as a copy of the original BaseAttachedProperty though the files are currently identical.  Should this be moved since its common to projects and the code is identical except for the class name changes? For now, BaseAttachedProperty.cs remains in Pharmacy2UApplication as a base page for the NoHistory attached property.  All page animations now inherit from BaseAnimationAttachedProperty, in the event this needs to be modified differently at some point in the future.

Moved all animation attached properties and animation helper items to the class library

All animation functionality appears to be working as originally with this refactoring at this time.