FInalized a dynamic visual status control in OrderItemControl.xaml that displays a different number of rectangles depending on the current status of order.  Also varies the colors depending on the status using a new value converter.
Retrieving additional information from the P2U_Orders table.

Added a custom routine to determine the status of an order regardless of what is stored in the Status string of P2U_Order.  While not ideal, this eliminates the need for a foreign key into the Statuses table and requires no changes in the existing database setup.

More bindings needed in the OrderItemsControl and the FullOrderInformation control.