More work on OrderItemControl underway.  Minor aesthetic changes made.
Able to now perform a secondary search of the zip codes table to match to the "zip" in the customer table in order to retrieve city, state, country, and county information.  A better way would be to create foreign keys in the database tables, but...  For now we use the databases as they are set up.

Still need to retrieve pharmacy, delivery company, and food / otc order items to finish the list.

This control should be reuseable by all of the different filter routines.

Need to pull a lot of the search queries out of AllOrdersPage into a more generic class to be used by all pages.