Continued work on AdminFoodEditControl.  Bindings now in place.  Save / Edit / Cancel functionality included.  Created an appropriate view model for each control.  Still need to intialize the viewmodel values with the actual data from the parent control.
Edit mode created for the control.  Will apply to other admin controls as necessary.  Refactoring for the controls possible?

Still need to read in all of the food items from the database, but the prototype template is functional.

Basic styling also added for textblocks, textboxes and buttons of the admin controls.  More to come...