Resolved issues with searching records.
Temporarily disconnected the real database data for a simple test model for searching.  Once the searching model is complete, we will reestablish the true data link.

Test case for simple string search established.  ESC once clears the field, ESC twice closes the search window.  RETURN executes the search of a filtered list.

Also cleaned up the number of times the database will be accessed (once on application load, and then only when a new record is received).

Instead, created an Items list and a FilteredItems list in the OrderListViewModel to handle the manipulations of the searching for now.

Basic formatting for search fields (a la AngelSix WPF32 tutorial) has been applied in Styles/Texts.XAML that utilizes tag information to display a mask over text boxes....

More restructuring of the various Order Screens will be incoming.