Incorporated an OrderViewModel object to control the database queries.
Incporated OrderViewModel into the IoC Dependency Inject in Pharmacy2UApplication.  This will allow us to control the number of times that the database needs to do a full read of the database.  This should significantly increase search speeds.

Database data is now read instead of the dummy data in the constructor.

Some minor graphical issues need to be addressed, but main functionality appears to be working.

Visual status indicator is now appearing again.