FoodItems and OrderItems now populate from the database instead of using the dummy data.
Worked on advanced search features and delivery status filters.  Multiple filters may be applied at once.  Some minor issues still with clearing and reloading the filtered items when a checkbox check is removed from the status filters.

Added a temporary link to the NewOrders Page that shows the basic printout template.  This will be removed and incorporated into the OrderItemListControl....maybe a slide in window or somesuch.

Changed OrderReadyForDelivery to OrderReadyForPickup to maintain consistency across the application.