Basic Popup window functionality implemented using the XAML Popup control.  Simple storyboard animations applied to IsOpen=true and IsOpen=false events (currently triggered by mouse clicks on buttons).
Still need to hook in scanning and monitoring of the database to this monitor.

Some work on RenderTransform and TranslateTransform still needed.  Does this need to be applied to the Canvas itself, or can we preset the canvas size in a control and then translate that control instead?  Or do we need a custom animation on loading like we did in the main Pharmacy2U application with pages and framework elements?